Separated cypress tests for mocha ones
