+++ src/server/.tests/mocha/e2e/applicationRun.spec.js
@@ -1,9 +1,9 @@
-const config = require('../../config.json')
+const config = require('../../../config.json')
 let suite
 describe('application startup', () => {
 	beforeEach('starting server instance for test', async () => {
 		suite = {}
-		suite.application = await require('../../init')({httpPort: config.devPropeties.httpTestPort})
+		suite.application = await require('../../../init')({httpPort: config.devPropeties.httpTestPort})
 	})
 
 	afterEach('clearing network connections after test', () => {
