+++ src/server/.tests/mocha/routes/controllers/register.spec.js
@@ -1,7 +1,7 @@
-const registerController = require('../../../routes/controllers/register')
+const registerController = require('../../../../routes/controllers/register')
 const assert = require('chai').assert
 const httpMocks = require('node-mocks-http')
-const responseMessages = require('../../../routes/controllers/utilities/responseMessages')
+const responseMessages = require('../../../../routes/controllers/utilities/responseMessages')
 const EventEmitter = require('events').EventEmitter
 let suite = {}
