+++ src/server/package.json
@@ -7,7 +7,7 @@
     "build": "webpack --config ./prod.webpack.config.js && cp ./frontEnd/src/theme/index.html ./public/",
     "build-dev": "webpack --config ./dev.webpack.config.js && cp ./frontEnd/src/theme/index.html ./public/",
     "start": "node ./index.js",
-    "test": "./node_modules/mocha/bin/mocha './.tests/**/*.spec.js'",
+    "test": "./node_modules/mocha/bin/mocha './.tests/mocha/**/*.spec.js'",
     "eslint": "node ./node_modules/eslint/bin/eslint.js --ignore-path ./.eslintignore ./",
     "eslint-auto-fix": "node ./node_modules/eslint/bin/eslint.js --ignore-path ./.eslintignore --fix",
     "generate-docs": "node ./node_modules/jsdoc/jsdoc -r ./ws-routes/",
@@ -18,12 +18,11 @@
   },
   "nyc": {
     "exclude": [
-      "**/*.spec.js",
       "out/",
       "frontEnd/",
       "public/",
       "runCypress.js",
-      "cypress/"
+      ".tests/"
     ],
     "reporter": [
       "text",
